+++ src/test/suite/runnerConfig.test.ts
@@ -54,14 +54,14 @@ describe('runnerConfig', async () => {
 		});
 	});
 	
-	describe('playwrightDebugOptions', () => {
-		it('playwrightDebugOptions test 1', async () => {
-			await conf.update('playwrightDebugOptions', {a:123});
-			assert.deepStrictEqual({a:123}, config.playwrightDebugOptions);
+	describe('playwrightEnvironmentVariables', () => {
+		it('playwrightEnvironmentVariables test 1', async () => {
+			await conf.update('playwrightEnvironmentVariables', ["AA=123"]);
+			assert.deepStrictEqual(["AA=123"], config.playwrightEnvironmentVariables);
 		});
-		it('playwrightDebugOptions test default', async () => {
-			await conf.update('playwrightDebugOptions', undefined);
-			assert.deepStrictEqual({}, config.playwrightDebugOptions);
+		it('playwrightEnvironmentVariables test default', async () => {
+			await conf.update('playwrightEnvironmentVariables', undefined);
+			assert.deepStrictEqual([], config.playwrightEnvironmentVariables);
 		});
 	});